Adapt library so that it distinguishes between numeral and number.
Also:
Adapt README.rst to match and add a remark about character digits.

Compute the value of a numeral more efficiently in numeral_to_number.

Use Python 3 '//' in case this is used in Python 3.

Check for base that exceed maximum allowed.

Adapt some tests appropriately, and add a few more.

Signed-off-by: mulhern <mulhern@cs.wisc.edu>